Added Input Manager
Added Input Manager and changed the ball hitting, spellcasting, and pausing game portions of code to instead use the Input Manager to clean up Update() functions.

Also changed the format of the spellBook variable in Spellcasting to a dictionary and changed the checkSpell() function accordingly to work as intended again.