Connected pickups & debuffs to spellcasting, added spellcasting UI, and fixed AI aiming system.
Spell pickup prefabs no longer immediately apply debuff, and instead add that spell to the player's spellbook.
Made it so that once a spell has been added to the spellbook, it won't spawn in on the court anymore.
Added a second spell.
Connected the spellcasting to the debuff system for the AI (when a spell is properly cast, it applies the debuff to the opponent AI).
Made UI for the spellcasting and spellbook that appears while the player is spellcasting (shows what the player's currently inputted spell address is, shows all spells within the spellbook, and as player is inputting spell the spellbook will update to only show spells that match the spell address the player is inputting (like a spell search function essentially)).
Fixed the AI aiming system so it's no longer hitting the net.