Fixed bounce checker and altered spellcasting
Bounce checker will now end the round if the ball bounces twice on the player's side of the court. This has also fixed the bugs regarding the AI hitting the net and the player being able to stay at the back of the court.

Altered spellcasting so that the spellbook is always open and spellcasting is always active.
Left shift now casts the spell, and the player can cancel their current inputted spell by pressing left control.
This also fixes the error regarding the game crashing when trying to spellcast after clicking 'Next Round'.